iOS 3.0.2
App icon image added.
Some bugs fixed.
There is a bug in Google Heat map and it desapears in a certain level of zoom-in.
The current location button has been updated.
Constratints added to slider, current location and the textfield. So they stay in the same position on all iOS devices with different display size.